Added dashboard and profile page
The dashboard page has been added. The user is able to create posts and the name of the user who created the post renders next to the post. The user's name is also a link that redirects to a profile page that currently displays their name and email. Validations for post creation are also working.