Added postLike method to join the likes and posts tables
I added another join statement and should have all the necessary join statements to achieve the desired functionality. Currently working through a bug where an error 404(Not Found) is returned when a user tries to create a post. The end points on the form and the route in the controller match, I still have been unable to determine the cause.